+++ src/components/grid/body.jsx
@@ -5,11 +5,11 @@ import GridRow from './row';
 
 export default class GridBody extends Component {
   @withProps
-  render({ columns, bodyWidth, data }) {
-    return <Container className="rx-grid-body">
+  render({ columns, width, bodyWidth, data, ...others }) {
+    return <Container className="rx-grid-body" style={{ width }}>
       <section className="rx-grid-view" style={{ width:bodyWidth }}>
         <div style={{ height:1 }} />
-        {data && data.map((record, rowIndex) => <GridRow columns={columns} record={record} rowIndex={rowIndex} />)}
+        {data && data.map((record, rowIndex) => <GridRow rowIndex={rowIndex} record={record} columns={columns} {...others} />)}
       </section>
     </Container>;
   }