+++ src/components/grid/header.jsx
@@ -1,16 +1,18 @@
 import React, { Component } from 'react';
 import withProps from '~/mixin/with-props';
+import { Checkbox } from '~/components/form';
 
 export default class GridHeader extends Component {
   @withProps
-  render({ columns, headerWidth }) {
-    return <section className="rx-grid-header">
+  render({ width, headerWidth, checkColumn, total, columns }) {
+    return <section className="rx-grid-header" style={{ width }}>
       <div className="rx-grid-header-container d-flex flex-row" style={{ width:headerWidth }}>
+        {checkColumn && <div className="rx-grid-column-header text-sm-center" style={{width:26}}>
+          {(checkColumn === 'multiple') && <Checkbox disabled={total === 0} />}
+        </div>}
         {columns && columns.map(col => {
-          const { text, className, style, ...others } = col;
-          return <div className={`rx-grid-column-header text-center ${className || ''}`} style={style} { ...others }>
-            {text || ''}
-          </div>
+          const { text = '', className = '', ...others } = col;
+          return <div className={`rx-grid-column-header text-center ${className}`} { ...others }>{text}</div>
         })}
       </div>
     </section>;