+++ src/data/model.js
@@ -1,12 +1,13 @@
 import Ext from '~/core/ext';
+import List from '~/core/list';
+import Map from '~/core/map';
 import Observable from '~/mixin/observable';
 
 export default class Model {
   constructor(data, store) {
-    Ext.extend(this, {
-      data,
-      store
-    });
+    this.data = data;
+    this.store = store;
+    this.fields = this.createFields((store && store.fields) ? store.fields : Object.keys(this.data));
     this.save();
   }
 
@@ -28,12 +29,35 @@ export default class Model {
 
   save() {
     this.phantom = Ext.isPrimitive(this.data) ? this.data : Ext.extend({}, this.data);
-    this.store && this.store.observable.call(this.store);
   }
 
   reject() {
     this.data = Ext.extend({}, this.phantom);
     this.save();
-    this.store && this.store.observable.call(this.store);
+  }
+
+  isModified(fieldName) {
+    if (fieldName) {
+      return !this.isEqual(this.fields[fieldName]);
+    }
+
+    return Map(this.fields).values().contains(field => !this.isEqual(field));
+  }
+
+  isEqual(field) {
+    return !field ? true :
+        (field.equals ? field.equals(this.data[field.name], this.phantom[field.name]) :
+            (this.data[field.name] === this.phantom[field.name]));
+  }
+
+  createFields(fields) {
+    return List(fields).reduce((fieldByName, field) => {
+      if (Ext.isObject(field)) {
+        fieldByName[field.name] = field;
+      } else {
+        fieldByName[field] = { name: field };
+      }
+      return fieldByName;
+    }, {});
   }
 }