+++ CHANGES.md
@@ -2,6 +2,7 @@
 > Jun 18, 2017
 
   * Provide `Dropdown` component
+  * Provide Grid Paging Toolbar
   * Support `editable` for `Grid`
   * Add function `Rext.generateSetter` to handle `setState` for each field in state
   * Fix issue in routes